chore(rswag-ui): Organise basic auth files
